Refactor SessionManagementTests.swift into separate test files
